requirements.txt
fastapi
uvicorn[standard]
pillow
torch
torchvision
opencv-python-headless
pandas
requests
ultralytics
seaborn
python-multipart